refactor(read): 重构排版与替换规则相关逻辑
1. 重构页眉页脚配置为统一排版配置页面
2. 分离正文与页眉页脚字体字号设置，新增独立页脚配置
3. 替换规则页面新增启用开关、生效替换与内容处理入口
4. 目录页面新增更新目录按钮
5. 修复标题字号计算逻辑，避免负值异常
6. 优化自定义页眉页脚渲染逻辑，支持独立配置页脚样式
7. 调整多语言字符串，统一术语与命名
