+++ app/src/main/java/io/legado/app/ui/book/read/page/provider/ChapterProvider.kt
@@ -1013,7 +1013,15 @@ object ChapterProvider {
         tPaint.typeface = titleFontTypeface
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P && ReadBookConfig.titleBold in 100..900)
             tPaint.setFontVariationSettings("'wght' ${ReadBookConfig.titleBold}")
-        tPaint.textSize = with(ReadBookConfig) { textSize + titleSize }.toFloat().spToPx()
+        tPaint.textSize = with(ReadBookConfig) {
+            if (titleSize >= 8) {
+                titleSize
+            } else if (titleSize > 0) {
+                textSize + titleSize
+            } else {
+                textSize
+            }
+        }.toFloat().spToPx()
         tPaint.isAntiAlias = true
         if (Build.VERSION.SDK_INT <= Build.VERSION_CODES.Q && ReadConfig.optimizeRender) {
             tPaint.isLinearText = true
