+++ app/src/main/res/values-zh-rCN/strings.xml
@@ -248,7 +248,8 @@
     <string name="import_book_source">导入书源</string>
     <string name="import_local">本地导入</string>
     <string name="import_on_line">网络导入</string>
-    <string name="replace_rule_title">替换净化</string>
+    <string name="replace_rule_enabled">启用替换净化</string>
+    <string name="replace_rule_title">替换规则</string>
     <string name="replace_rule_edit">替换规则编辑</string>
     <string name="replace_rule">替换规则</string>
     <string name="replace_to">替换为</string>
@@ -375,6 +376,7 @@
     <string name="fonts_folder">请将字体文件放到存储根目录 Fonts 文件夹下后重新选择</string>
     <string name="default_font">默认字体</string>
     <string name="select_font">选择字体</string>
+    <string name="font">字体</string>
     <string name="text_size">字号</string>
     <string name="line_size">行距</string>
     <string name="paragraph_size">段距</string>
@@ -900,7 +902,13 @@
     <string name="read_config_regex_desc">根据正则表达式更换字体与颜色</string>
     <string name="read_config_header_footer_desc">配置页眉页脚模块、显示规则和颜色</string>
     <string name="read_config_header_footer_padding_desc">调整页眉页脚上下左右边距</string>
-    <string name="read_config_divider_line">分割线</string>
+    <string name="read_config_divider_line">分隔线</string>
+    <string name="read_config_display">显示</string>
+    <string name="read_config_spacing">间距</string>
+    <string name="read_config_content">内容</string>
+    <string name="read_config_image">图片</string>
+    <string name="header_font_size">页眉字号</string>
+    <string name="apply_header_style">应用页眉字体样式设置</string>
     <string name="group_name">分组名称</string>
     <string name="note_content">备注</string>
     <string name="replace_enable_default_t">默认启用替换净化</string>
@@ -965,9 +973,9 @@
     <string name="next_page">下一页</string>
     <string name="prev_page">上一页</string>
     <string name="non_action">无操作</string>
-    <string name="body_title">正文标题</string>
+    <string name="body_title">标题</string>
     <string name="show_hide">显示/隐藏</string>
-    <string name="header_footer">信息</string>
+    <string name="header_footer">排版</string>
     <string name="rule_subscription">规则</string>
     <string name="rule_sub_empty_msg">还没有规则订阅！</string>
     <string name="get_book_progress">拉取云端进度</string>
@@ -1018,7 +1026,7 @@
     <string name="use_default_cover">总是使用默认封面</string>
     <string name="use_default_cover_s">总是显示默认封面（不显示网络封面）</string>
     <string name="title_font_size">字号</string>
-    <string name="title_color">标题颜色</string>
+    <string name="title_color">颜色</string>
     <string name="reset_to_body_color">重置为正文颜色</string>
     <string name="title_margin_top">上边距</string>
     <string name="title_margin_bottom">下边距</string>
@@ -1685,7 +1693,7 @@
     <string name="expand_volume">展开所有卷</string>
     <string name="custom">自定义</string>
     <string name="global_setting">全局设置</string>
-    <string name="text_typeface">字体样式</string>
+    <string name="text_typeface">字体</string>
     <string name="shadow_typeface">阴影样式</string>
     <string name="long_click_to_custom">长按自定义</string>
     <string name="disable_manga_scroll_animation">禁用翻页动画</string>
@@ -1814,9 +1822,6 @@
     <string name="dashed_border">虚线</string>
     <string name="header_color">页眉</string>
     <string name="footer_color">页脚</string>
-    <string name="header_font">设置字体</string>
-    <string name="footer_font">页脚字体</string>
-    <string name="header_font_size">设置字号</string>
     <string name="source_tab_start">启动</string>
     <string name="r_startHtml">启动页html文本</string>
     <string name="r_startStyle">启动页style文本</string>
