+++ app/src/main/res/values-zh-rTW/strings.xml
@@ -200,7 +200,8 @@
     <string name="import_book_source">匯入書源</string>
     <string name="import_local">本機匯入</string>
     <string name="import_on_line">網路匯入</string>
-    <string name="replace_rule_title">取代淨化</string>
+    <string name="replace_rule_enabled">啟用取代淨化</string>
+    <string name="replace_rule_title">取代規則</string>
     <string name="replace_rule_edit">取代規則編輯</string>
     <string name="replace_rule">取代規則</string>
     <string name="replace_to">取代為</string>
@@ -323,6 +324,7 @@
     <string name="fonts_folder">請將字體檔案放到SD根目錄Fonts資料夾下重新選擇</string>
     <string name="default_font">預設字體</string>
     <string name="select_font">選擇字體</string>
+    <string name="font">字體</string>
     <string name="text_size">字號</string>
     <string name="line_size">行距</string>
     <string name="paragraph_size">段距</string>
@@ -842,9 +844,9 @@
     <string name="next_page">下一頁</string>
     <string name="prev_page">上一頁</string>
     <string name="non_action">無操作</string>
-    <string name="body_title">正文標題</string>
+    <string name="body_title">標題</string>
     <string name="show_hide">顯示/隱藏</string>
-    <string name="header_footer">頁首<![CDATA[&]]>頁尾</string>
+    <string name="header_footer">排版</string>
     <string name="rule_subscription">規則訂閱</string>
     <string name="rule_sub_empty_msg">新增大佬們提供的規則匯入地址\n新增後點擊可匯入規則</string>
     <string name="get_book_progress">拉取雲端進度</string>
@@ -1319,9 +1321,6 @@
     <string name="dashed_border">虚线</string>
     <string name="header_color">页眉</string>
     <string name="footer_color">页脚</string>
-    <string name="header_font">设置字体</string>
-    <string name="footer_font">页脚字体</string>
-    <string name="header_font_size">设置字号</string>
     <string name="source_tab_start">启动</string>
     <string name="r_startHtml">启动页html文本</string>
     <string name="r_startStyle">启动页style文本</string>
@@ -1521,6 +1520,12 @@
     <string name="read_config_header_footer_desc">設定頁首/頁尾模組、可見性和顏色</string>
     <string name="read_config_header_footer_padding_desc">調整頁首和頁尾的上、下、左、右邊距</string>
     <string name="read_config_divider_line">分隔線</string>
+    <string name="read_config_display">顯示</string>
+    <string name="read_config_spacing">間距</string>
+    <string name="read_config_content">內容</string>
+    <string name="read_config_image">圖片</string>
+    <string name="header_font_size">頁首字號</string>
+    <string name="apply_header_style">應用頁首字型樣式設定</string>
     <string name="brightness_bar_off">關閉</string>
     <string name="brightness_bar_horizontal">水平</string>
     <string name="brightness_bar_vertical">垂直</string>
@@ -1531,7 +1536,7 @@
     <string name="image_style_full">全螢幕</string>
     <string name="image_style_text">文字吸附</string>
     <string name="image_style_single">單頁</string>
-    <string name="title_color">標題顏色</string>
+    <string name="title_color">顏色</string>
     <string name="reset_to_body_color">重置為正文顏色</string>
     <string name="export_file_name_template_hint">{name} - {author}</string>
     <string name="export_file_name_template_help">可用變數：{name} 書名、{author} 作者、{group} 分組、{source} 來源、{remark} 備註。
@@ -1831,7 +1836,7 @@
     <string name="expand_volume">展開所有卷</string>
     <string name="custom">自訂</string>
     <string name="global_setting">全域設定</string>
-    <string name="text_typeface">字體樣式</string>
+    <string name="text_typeface">字體</string>
     <string name="shadow_typeface">陰影樣式</string>
     <string name="long_click_to_custom">長按自訂</string>
     <string name="disable_manga_scroll_animation">停用翻頁動畫</string>
